+++ backend/src/main/java/org/yellowcat/backend/product/address/AddressService.java
@@ -1,4 +1,4 @@
-package org.yellowcat.backend.address;
+package org.yellowcat.backend.product.address;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -6,10 +6,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.yellowcat.backend.user.AppUser;
 import org.yellowcat.backend.user.AppUserService;
-import org.yellowcat.backend.address.dto.AddressesDTO;
+import org.yellowcat.backend.product.address.dto.AddressesDTO;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class AddressService {
