+++ backend/src/main/java/org/yellowcat/backend/product/order/OrderRepository.java
@@ -1,0 +1,9 @@
+package org.yellowcat.backend.product.order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.Optional;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+}
