+++ backend/src/main/java/org/yellowcat/backend/product/orderItem/OrderItemRepository.java
@@ -1,0 +1,9 @@
+package org.yellowcat.backend.product.orderItem;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.List;
+
+@Repository
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+}
