+++ backend/src/main/java/org/yellowcat/backend/product/payment/paymentRepository.java
@@ -1,0 +1,8 @@
+package org.yellowcat.backend.product.payment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface paymentRepository extends JpaRepository <Payment, Integer> {
+}
