+++ backend/src/main/java/org/yellowcat/backend/product/shipment/ShipmentRepository.java
@@ -1,0 +1,8 @@
+package org.yellowcat.backend.product.shipment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
+}
